files/conversion/cifar10/requirements.txt
numpy==1.17.4
Pillow==7.0.0
